backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Google Gemini
google-generativeai==0.3.2

# LangChain
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16

# ChromaDB
chromadb==0.4.22

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1



# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
